+++ tools/flexbot-cli/flexbot.go
@@ -19,7 +19,7 @@ import (
 )
 
 const (
-	version = "1.7.4"
+	version = "1.7.5"
 )
 
 type OperationResult interface {